+++ BitcoinTransfer/ViewModels/LastTransactionModel.cs
@@ -10,7 +10,7 @@ namespace BitcoinTransfer.ViewModels
 
         public string Address { get; set; }
 
-        public double Amount { get; set; }
+        public decimal Amount { get; set; }
 
         public int Confirmation { get; set; }
     }
